Fixes shadows and outline scaling
Put all mana symbol rendering together so that we don't render them twice (we were before for shadow). Also makes more explicit check for outlines before changing how mana symbols are handled vs how they were. Updates the scaling on outlines to mirror how the text is outlined. Also changes the outline on set symbols to use the same type of rendering as the mana symbol outline and fixes the scaling of it as well so that it scales the same as the outline on everything else. This should help unify the outline for everything.