Japan Showcase Tweaks
Fixes Color On the Multi Frame so it uses the official color. Adds Black (Alt) frame and margin with the purple frame color from VOW showcase frames. Adds a PT Bottom Border for use when using the PT Box on it's own. Removes the Mana Backing from the frame elements since we now have outlines natively. Removes Excess PT Box (C) that was not being used.